refactor: move example to separate folder
